update to use larger numbers for blockGasLimit and trasactionFeeRates

+++ contracts/blockGasLimit-dgp.sol
@@ -1,7 +1,7 @@
 pragma solidity 0.5.8;
 
 contract blockGasLimit {
-    uint32[1] _blockGasLimit = [
+    uint256[1] _blockGasLimit = [
         40000000 //default block gas limit
     ];
     function getBlockGasLimit() public view returns (uint32[1] memory) {
